+++ AlexLeeTakeHomeWeb/Models/PurchaseDetailItems/SearchModel.cs
@@ -1,9 +1,9 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
-namespace AlexLeeTakeHomeWeb.Models;
+namespace AlexLeeTakeHomeWeb.Models.PurchaseDetailItems;
 
-public class PurchaseDetailItemSearchModel
+public class SearchModel
 {
 	[DisplayName("Purchase Order Number")]
 	public string? PurchaseOrderNumber { get; set; } = null!;
