+++ AlexLeeTakeHomeWeb/Views/PurchaseDetailItems/Create.cshtml
@@ -1,4 +1,5 @@
-﻿@model PurchaseDetailItemCreateModel
+﻿@using AlexLeeTakeHomeWeb.Models.PurchaseDetailItems
+@model CreateModel
 
 @{
     ViewData["Title"] = "Create";
