+++ AlexLeeTakeHomeWeb/Views/PurchaseDetailItems/Index.cshtml
@@ -1,4 +1,5 @@
-﻿@model PurchaseDetailItemIndexModel
+﻿@using AlexLeeTakeHomeWeb.Models.PurchaseDetailItems
+@model IndexModel
 
 @{
     ViewData["Title"] = "Purchase Detail Items";
