+++ AlexLeeTakeHomeWeb/Views/PurchaseDetailItems/_Search.cshtml
@@ -1,4 +1,5 @@
-﻿@model PurchaseDetailItemSearchModel
+﻿@using AlexLeeTakeHomeWeb.Models.PurchaseDetailItems
+@model SearchModel
 
 @{
 }
